split das_import into a module and a script
